pyproject.toml
[project]
name = "timetable-sync-api"
description = "API wrapper powering timetable-sync"
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "nova" }]
dependencies = [
    "aiohttp>=3.13.3",
    "icalendar>=6.3.2",
    "msgspec>=0.20.0",
    "rapidfuzz>=3.14.3",
    "valkey-glide>=2.2.6",
]

[project.urls]
Repository = "https://github.com/novanai/timetable-sync-api"

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "ruff>=0.14.14",
    "types-icalendar>=6.3.2.20251222",
    "types-redis>=4.6.0.20241004",
    "uv>=0.9.26",
]
nox = [
    "nox>=2025.11.12",
]

[tool.ruff]
include = ["./timetable/*.py", "noxfile.py"]
target-version = "py313"

[tool.ruff.format]
docstring-code-format = true
line-ending = "lf"

[tool.ruff.lint]
select = [
    "F", # Pyflakes
    "E", # Error (pycodestyle)
    "W", # Warning (pycodestyle)
    "I", # isort
    "N", # pep8-naming
    "ANN", # flake8-annotations
    "ASYNC", # flake8-async
    "A", # flake8-builtins
    "COM", # flake8-commas
    "C4", # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "ICN", # flake8-import-conventions
    "Q", # flake8-quotes
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "TC", # flake8-type-checking
    "ARG", # flake8-unused-arguments
    "ERA", # eradicate
    "PL", # Pylint
    "PERF", # Perflint
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501", # line-too-long
    "PLR2004", # magic-value-comparison
    "PLR0913", # too-many-arguments
    "COM812", # missing-trailing-comma
    "ANN401", # any-type
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.uv]
required-version = "~=0.9"

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
